Fix no_sync case
For no_sync session, we need to mark param used as long as it is ever used in any iteration during the session. In that way, we can get the correct params which are not used locally in the end of the session and then reduce the used maps to figure out the globally unused params during the entire no_sync session.