Add distributed data parallel benchmark tool
The need for this tool was motivated by #28883. In the past, we have
done ad-hoc benchmarking, but it's time for something more structured.

It would be nice to add more model architectures so that we can get a
full picture of the performance impact of a code change simply by
running this suite a few times.

ghstack-source-id: a122c2057a902e383c41d7154bfa1ce1f3636076
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30660