Change python/cpp docs CI to use a CPU-only image
Previously, the docs CI depended on our CUDA xenial py3 build. This
meant that the turnaround time to get signal for docs was very slow
(I've seen builds that go as much as 3 hours).

Fortunately, the docs CI do not (and should not!) rely on CUDA. This
PR changes it so that the docs CI runs on a CPU-only machine.

Fixes https://github.com/pytorch/pytorch/issues/29995

Test Plan:
- Check CI status on this PR by reading logs for the python and cpp docs
builds.

ghstack-source-id: 35605e5dd6491b273b37448cc30a90b8e2d4a425
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30487