[BC-BREAKING] change index_select scalar_check to retain dimensionality of input.
The index_select documentaiton reads:
"The returned tensor has the same number of dimensions as the original tensor (input)."

But the implementation would return a 0-dimensional tensor iff both the input and index were 0-dimensional.
This change makes it so we retuan a 0-dimensional tensor iff the input is 0-dimensional.

Restacked version of: https://github.com/pytorch/pytorch/pull/30502

ghstack-source-id: 16862bdb74f7ac163a4bc36b7bfcafbf082730b9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30790