[jit] Enable IValue to hold a PyObject
Summary:

This PR enables IValue to be able to hold a pure PyObject by adding a
new enum tag, a new jit_type to denote PyObject existance in IValue and
the JIT type system. We don't and not plan to expose this to user.

This is the basic piece that enable ivalue to be adopted broader like
making RRef always hold IValue, it might also simplify some compiler
logic

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 72377e13591970c35419ceb1704384fe92dd9bba
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30136